backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
pymongo==4.6.0
pinecone-client==2.2.4
openai==1.3.8
anthropic==0.7.8
google-generativeai==0.3.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-anthropic==0.0.1
sentence-transformers==2.2.2
googletrans==4.0.2
spacy==3.7.2
python-socketio==5.10.0
python-telegram-bot==20.7
twilio==8.11.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
Pillow==10.1.0
PyPDF2==3.0.1
unstructured==0.11.6
celery==5.3.4
flower==2.0.1
sentry-sdk[fastapi]==1.38.0
prometheus-fastapi-instrumentator==6.1.0

# Additional dependencies for RAG system
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0
motor==3.3.2
